Add basic pundit settings

+++ app/controllers/places_controller.rb
@@ -1,13 +1,16 @@
 class PlacesController < ApplicationController
+  after_action :verify_authorized, only: [:create, :destroy, :update]
 
   def create
     place = current_user.places.build(place_params)
+    authorize(place)
     place.save
     redirect_to current_user
   end
 
   def destroy
     place = Place.find(params[:id])
+    authorize(place)
     place.destroy
     respond_to do |format| 
       format.html { redirect_to place.user }
@@ -24,6 +27,7 @@ class PlacesController < ApplicationController
 
   def update
     @place = Place.find(params[:id])
+    authorize(@place)
     @place.update_attributes(place_params)
     redirect_to @place.user
   end
